requirements.txt
fastapi==0.111.0
pillow==10.3.0
uvicorn==0.30.1
